Don't add BUILD_NUMBER to version for platform builds
We're removing differences between two builds that have the same source.
These builds will have different BUILD_NUMBERs though, so it shouldn't
be embedded in the binary.

For eng builds, using the current date forces the APK to be rebuilt
every time the makefiles are read. Remove this for platform builds, but
keep it if building unbundled.

Bug: 24201956
Change-Id: I8c0ad93e817c74c69b8d8156d4727952b08024d4
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>